+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/db/derived/DbDerivedDataRepository.kt
@@ -1,11 +1,13 @@
 package cash.z.ecc.android.sdk.internal.db.derived
 
+import cash.z.ecc.android.sdk.internal.model.DbBlock
 import cash.z.ecc.android.sdk.internal.model.DbTransactionOverview
 import cash.z.ecc.android.sdk.internal.model.EncodedTransaction
 import cash.z.ecc.android.sdk.internal.repository.DerivedDataRepository
 import cash.z.ecc.android.sdk.model.AccountUuid
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.FirstClassByteArray
+import cash.z.ecc.android.sdk.model.TransactionId
 import cash.z.ecc.android.sdk.model.TransactionRecipient
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -52,12 +54,26 @@ internal class DbDerivedDataRepository(
     override val allTransactions: Flow<List<DbTransactionOverview>>
         get() = invalidatingFlow.map { derivedDataDb.allTransactionView.getAllTransactions().toList() }
 
-    override fun getOutputProperties(transactionId: FirstClassByteArray) =
+    override fun getOutputProperties(transactionId: TransactionId) =
         derivedDataDb.txOutputsView
-            .getOutputProperties(transactionId)
+            .getOutputProperties(transactionId.value)
 
-    override fun getRecipients(transactionId: FirstClassByteArray): Flow<TransactionRecipient> {
-        return derivedDataDb.txOutputsView.getRecipients(transactionId)
+    override fun getTransactionsByMemoSubstring(query: String): Flow<List<TransactionId>> =
+        derivedDataDb
+            .txOutputsView
+            .getTransactionsByMemoSubstring(query)
+            .map { listOfFirstClassByteArray ->
+                listOfFirstClassByteArray.map {
+                    TransactionId(it)
+                }
+            }
+
+    override fun getRecipients(transactionId: TransactionId): Flow<TransactionRecipient> {
+        return derivedDataDb.txOutputsView.getRecipients(transactionId.value)
+    }
+
+    override suspend fun findBlockByHeight(blockHeight: BlockHeight): DbBlock? {
+        return derivedDataDb.blockTable.findBlockByExpiryHeight(blockHeight)
     }
 
     override suspend fun close() {