+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/model/DbBlock.kt
@@ -3,9 +3,8 @@ package cash.z.ecc.android.sdk.internal.model
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.FirstClassByteArray
 
-internal data class Block(
+internal data class DbBlock(
     val height: BlockHeight,
     val hash: FirstClassByteArray,
-    val time: Int,
-    val saplingTree: FirstClassByteArray
+    val blockTimeEpochSeconds: Long,
 )