+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `Synchronizer.getRecipients` will correctly return an address for an account existing in database
 
+### Added
+- `Synchronizer.getTransactionsByMemoSubstring()` has been added
+- `TransactionId` object has been added and used instead of `FirstClassByteArray` in `TransactionOverview` and 
+  `PendingTransaction` model classes
+
+### Changed
+- `Synchronizer.getTransactions(accountUuid)` and `Synchronizer.transactions` now internally fill in 
+  `TransactionOverview.blockTimeEpochSeconds` based on the related block time 
+- `Synchronizer.transactions` has been renamed to `Synchronizer.allTransactions` to emphasize the fact the API 
+  returns transactions for all the wallet accounts 
+
 ## [2.2.7] - 2024-12-18
 
 ### Added